+++ package.json
@@ -54,6 +54,7 @@
     "helmet": "^3.22.0",
     "knex": "^0.21.0",
     "knex-cleaner": "^1.3.0",
+    "moment": "^2.25.3",
     "morgan": "^1.10.0",
     "pg": "^8.0.2",
     "supertest": "^4.0.2"
