+++ resources/visualizations/cases/cases.controllers.js
@@ -1,4 +1,4 @@
-const { queryMapData, queryByCountryData } = require("./cases.model");
+const { queryMapData } = require("./cases.model");
 const DatabaseError = require("../../../server/middleware/DatabaseError");
 
 const getVisualizationData = async (_req, res, next) => {
@@ -19,21 +19,6 @@ const getVisualizationData = async (_req, res, next) => {
   }
 };
 
-// Get Data for confirmed cases by country for Racing Chart Visualization
-const getConfirmedCasesData = async (_req, res, next) => {
-  try {
-    const confirmedData = await queryByCountryData();
-    res.status(200).json({
-      confirmed: confirmedData,
-    });
-  } catch (err) {
-    next(
-      new DatabaseError({
-        message: "Cannot retrieve cases",
-        dbMessage: err,
-      })
-    );
-  }
+module.exports = {
+  getVisualizationData,
 };
-
-module.exports = { getVisualizationData, getConfirmedCasesData };
